add: utils.ts on frontend and delete code not need it anymore
