+++ .eslintrc.json
@@ -24,6 +24,7 @@
         "prettier"
     ],
     "rules": {
-        "prettier/prettier": "error"
+        "prettier/prettier": "error",
+        "no-console": "warn"
     }
 }
