+++ src/render/components/Main/index.tsx
@@ -2,15 +2,15 @@
  Author: Johan Suarez Largo
  */
 // react
-import React, { useEffect, useState } from 'react';
-import { IpcRendererEvent } from 'electron/main';
+import React from 'react';
 import ListReference from '../ListReference';
 import TextBox from '../TextBox';
 import { ReferenceData } from '../../../main/DataStorage';
+import { parseReferenceToText } from '../../utils';
 import copyImg from '../../assets/img/content_copy_black_24dp.svg';
 import offLineImg from '../../assets/img/cloud_off_black_24dp.svg';
 import onLineImg from '../../assets/img/cloud_done_black_24dp.svg';
-
+import useMain from './useMain';
 //Interfaces
 
 export interface ListReferences {
@@ -20,65 +20,18 @@ export interface ListReferences {
 // Components
 
 function Main(): JSX.Element {
-  const [listReferences, setListReferences] = useState<ListReferences>({});
-
-  const addReference = (event: IpcRendererEvent, reference: ReferenceData) => {
-    setListReferences((prevList) => {
-      const newList = { ...prevList };
-      newList[reference.id] = reference;
-      return newList;
-    });
-  };
-
-  const dropReference = (id: string) => {
-    setListReferences((prevList) => {
-      const newList = { ...prevList };
-      delete newList[id];
-      return newList;
-    });
-  };
-
-  const onChangeReference = (
-    event: IpcRendererEvent,
-    action: string,
-    id: string,
-    result: ReferenceData | boolean
-  ): void => {
-    if (result) {
-      if (action === 'update') {
-        window.electron.dataAPI.getReference(id);
-      } else if (action === 'delete') {
-        dropReference(id);
-      } else if (action === 'partial-update') {
-        window.electron.dataAPI.getReference(id);
-      }
-    }
-  };
-
-  useEffect(() => {
-    // ICP communication
-    window.electron.onEventsAPI.onProcessURLReady(addReference);
-    window.electron.onEventsAPI.onReferenceReady(addReference);
-    window.electron.onEventsAPI.onChangeReference(onChangeReference);
-  }, []);
+  const listReferences = useMain();
 
   const copyClipBoard = (): void => {
     let text = '';
     for (const index in listReferences) {
       if (listReferences[index].status === 'ready') {
-        text = text + getRefText(listReferences[index]) + '\n';
+        text = text + parseReferenceToText(listReferences[index]) + '\n';
       }
     }
     window.electron.dataAPI.copyClipBoard(text);
   };
 
-  const getRefText = (reference: ReferenceData): string => {
-    if (reference.authorName === reference.authorSurname) {
-      return `${reference.authorSurname}. (${reference.yearPublish}). ${reference.title} Retrieved from <${reference.URL}>`;
-    } else {
-      return `${reference.authorSurname}, ${reference.authorName}. (${reference.yearPublish}). ${reference.title} Retrieved from <${reference.URL}>`;
-    }
-  };
   return (
     <main>
       <div className="container">
