+++ src/render/components/Reference.tsx
@@ -1,39 +1,33 @@
 import React from 'react';
 import { ReferenceData } from '../../main/DataStorage';
+import { parseReferenceToText } from '../utils';
 import copyImg from '../assets/img/content_copy_white_24dp.svg';
 import editImg from '../assets/img/edit_black_24dp.svg';
 import '../assets/styles/reference.css';
 
 interface IProps {
   reference?: ReferenceData;
-  index?: number;
 }
 
-const Reference = ({ reference, index }: IProps): JSX.Element => {
+const Reference = ({ reference }: IProps): JSX.Element => {
   const enableEdition = (id: string) => {
     window.electron.dataAPI.partialUpdateReference(id, { status: 'editing' });
   };
-  const copyClipBoard = (text: string) => {
+  const copyClipBoard = (reference: ReferenceData) => {
+    const text = parseReferenceToText(reference);
     window.electron.dataAPI.copyClipBoard(text);
   };
 
-  const getRefText = (): string => {
-    if (reference.authorName === reference.authorSurname) {
-      return `${reference.authorSurname}. (${reference.yearPublish}). ${reference.title} Retrieved from <${reference.URL}>`;
-    } else {
-      return `${reference.authorSurname}, ${reference.authorName}. (${reference.yearPublish}). ${reference.title} Retrieved from <${reference.URL}>`;
-    }
-  };
   return (
     <div className="row valign-wrapper">
       <div className="col s9">
-        <p id={index.toString()}>{getRefText()}</p>
+        <p id={reference.id}>{parseReferenceToText(reference)}</p>
       </div>
       <div className="col s2 offset-s1 ">
         <button
           className="btn-floating btn-container green accent-3"
           onClick={() => {
-            copyClipBoard(getRefText());
+            copyClipBoard(reference);
           }}
         >
           <img className="image-done" src={copyImg} alt="good" />
